73933 Create function for copying data
