+++ src/AplosConnector.Common/Storage/AzureTableStorageAbstract.cs
@@ -4,7 +4,6 @@ namespace AplosConnector.Common.Storage
 {
     public abstract class AzureTableStorageAbstract
     {
-        //TODO set back to protected after migration
         public readonly TableClient TableClient;
 
         protected AzureTableStorageAbstract(TableClient tableClient)
